requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
kafka-python>=2.0.0
mlflow>=2.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.0.0
jupyter>=1.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0
httpx>=0.24.0
pytest>=8.0.0
 
# PyTorch is optional. When installed the LSTM Autoencoder is used
# instead of the sklearn MLP Autoencoder fallback.
# Install separately: https://pytorch.org/get-started/locally/
# torch>=2.0.0